+++ docs/pages/integration-test/masonry.js
@@ -22,27 +22,35 @@ export default function TestPage(): Node {
   const router = useRouter();
   const {
     constrained,
+    externalCache,
     finiteLength,
     flexible,
     manualFetch,
+    noScroll,
     offsetTop,
     scrollContainer,
     virtualize,
+    virtualBoundsTop,
+    virtualBoundsBottom,
   } = router.query;
 
   return (
     <ColorSchemeProvider colorScheme="light">
       <MasonryContainer
         constrained={booleanize(constrained)}
+        externalCache={booleanize(externalCache)}
         finiteLength={booleanize(finiteLength)}
         flexible={booleanize(flexible)}
         initialItems={generateExampleItems({ name: 'InitialPin' })}
         manualFetch={booleanize(manualFetch)}
         MasonryComponent={Masonry}
         measurementStore={measurementStore}
+        noScroll={booleanize(noScroll)}
         offsetTop={offsetTop}
         scrollContainer={booleanize(scrollContainer)}
         virtualize={booleanize(virtualize)}
+        virtualBoundsTop={virtualBoundsTop}
+        virtualBoundsBottom={virtualBoundsBottom}
       />
     </ColorSchemeProvider>
   );