+++ docs/pages/visual-test/Toast.js
@@ -7,14 +7,16 @@ export default function Snapshot(): Node {
     <Toast
       primaryAction={{ accessibilityLabel: 'Test', label: 'Undo', size: 'lg' }}
       text="Home decor"
-      thumbnail={
-        <Image
-          alt="Modern ceramic vase pin."
-          naturalHeight={564}
-          naturalWidth={564}
-          src="https://i.ibb.co/Lx54BCT/stock1.jpg"
-        />
-      }
+      thumbnail={{
+        image: (
+          <Image
+            alt="Modern ceramic vase pin."
+            naturalHeight={564}
+            naturalWidth={564}
+            src="https://i.ibb.co/Lx54BCT/stock1.jpg"
+          />
+        ),
+      }}
     />
   );
 }