+++ docs/pages/web/switch.js
@@ -1,40 +1,27 @@
 // @flow strict
 import { type Node } from 'react';
+import AccessibilitySection from '../../docs-components/AccessibilitySection.js';
+import docgen, { type DocGen } from '../../docs-components/docgen.js';
 import GeneratedPropTable from '../../docs-components/GeneratedPropTable.js';
 import MainSection from '../../docs-components/MainSection.js';
 import Page from '../../docs-components/Page.js';
 import PageHeader from '../../docs-components/PageHeader.js';
-import docgen, { type DocGen } from '../../docs-components/docgen.js';
 import QualityChecklist from '../../docs-components/QualityChecklist.js';
-import AccessibilitySection from '../../docs-components/AccessibilitySection.js';
+import SandpackExample from '../../docs-components/SandpackExample.js';
+import main from '../../examples/switch/main.js';
+import doLabel from '../../examples/switch/doLabel.js';
+import dontTruncate from '../../examples/switch/dontTruncate.js';
+import doDisabled from '../../examples/switch/doDisabled.js';
+import dontAlternative from '../../examples/switch/dontAlternative.js';
+import variantLabel from '../../examples/switch/variantLabel.js';
+import variantCombinations from '../../examples/switch/variantCombinations.js';
 
 export default function DocsPage({ generatedDocGen }: {| generatedDocGen: DocGen |}): Node {
   return (
-    <Page title="Switch">
-      <PageHeader
-        name="Switch"
-        description={generatedDocGen?.description}
-        defaultCode={`
-function SwitchExample() {
-  const [switched, setSwitched] = React.useState(false);
-
-  return (
-    <Box display="flex" alignItems="center">
-      <Box paddingX={2}>
-        <Label htmlFor="introExample">
-          <Text>Airplane mode</Text>
-        </Label>
-      </Box>
-      <Switch
-        onChange={() => setSwitched(!switched)}
-        id="introExample"
-        switched={switched}
-      />
-    </Box>
-  );
-}
-      `}
-      />
+    <Page title={generatedDocGen?.displayName}>
+      <PageHeader name={generatedDocGen?.displayName} description={generatedDocGen?.description}>
+        <SandpackExample code={main} name="Primary Switch example" hideEditor previewHeight={150} />
+      </PageHeader>
 
       <GeneratedPropTable generatedDocGen={generatedDocGen} />
 
@@ -59,117 +46,53 @@ function SwitchExample() {
           />
         </MainSection.Subsection>
       </MainSection>
+
       <MainSection name="Best practices">
         <MainSection.Subsection columns={2}>
           <MainSection.Card
             cardSize="md"
             type="do"
             description="Use a label to give Switch context when possible."
-            defaultCode={`
-function SwitchExample() {
-  const [switched, setSwitched] = React.useState(true);
-
-  return (
-    <Flex gap={{ row: 2, column: 0 }} alignItems="center" width="100%" maxWidth={300}>
-      <Flex.Item flex="grow">
-        <Label htmlFor="makesecret">
-          <Text weight="bold">Make this board secret</Text>
-        </Label>
-      </Flex.Item>
-      <Switch
-        onChange={() => setSwitched(!switched)}
-        id="makesecret"
-        switched={switched}
-      />
-    </Flex>
-  );
-}
-`}
+            sandpackExample={<SandpackExample code={doLabel} name="Do - label" hideEditor />}
           />
           <MainSection.Card
             cardSize="md"
             type="don't"
-            description="Truncate label text. Instead, allow it to wrap to form another line. "
-            defaultCode={`
-function SwitchExample() {
-  const [switched, setSwitched] = React.useState(false);
-
-  return (
-    <Flex gap={{ row: 2, column: 0 }} alignItems="center">
-      <Switch
-        onChange={() => setSwitched(!switched)}
-        id="setboard"
-        switched={switched}
-      />
-      <Label htmlFor="setboard">
-        <Text>Set board to...</Text>
-      </Label>
-    </Flex>
-  );
-}
-`}
+            description="Truncate label text. Instead, allow it to wrap to form another line."
+            sandpackExample={
+              <SandpackExample
+                code={dontTruncate}
+                name="Don't - Truncate"
+                hideControls
+                hideEditor
+              />
+            }
           />
         </MainSection.Subsection>
+
         <MainSection.Subsection columns={2}>
           <MainSection.Card
             cardSize="md"
             type="do"
             description="Communicate why a switch is disabled and how to enable it if possible."
-            defaultCode={`
-<Flex gap={{ row: 2, column: 0 }} alignItems="center">
-  <Switch
-    onChange={() => {}}
-    id="showsecretboards"
-    switched={false}
-    disabled={true}
-  />
-  <Label htmlFor="showsecretboards">
-    <Flex direction="column" gap={{ column: 1, row: 0 }}>
-      <Text>Show secret boards</Text>
-      <Text size="100" color="subtle">You don't have any secret boards</Text>
-    </Flex>
-  </Label>
-</Flex>
-`}
+            sandpackExample={<SandpackExample code={doDisabled} name="Do - Disabled" hideEditor />}
           />
           <MainSection.Card
             cardSize="md"
             type="don't"
             description="Use alternative styling to represent the functionality of a switch. Use Switch instead."
-            defaultCode={`
-<Flex direction="column" gap={{ column: 2, row: 0 }} width="100%" maxWidth={300}>
-  <Heading size="300" accessibilityLevel={3}>On Pinterest notifications</Heading>
-  <Flex direction="column" gap={{ column: 4, row: 0 }}>
-    <Flex alignItems="center">
-      <Flex.Item flex="grow">
-        <Text>Comments</Text>
-      </Flex.Item>
-      <Icon icon="check-circle" color="default" accessibilityLabel="checked circle" />
-    </Flex>
-    <Flex alignItems="center">
-      <Flex.Item flex="grow">
-        <Text>Mentions</Text>
-      </Flex.Item>
-      <Icon icon="check-circle" color="default" accessibilityLabel="checked circle" />
-    </Flex>
-    <Flex alignItems="center">
-      <Flex.Item flex="grow">
-        <Text>Views</Text>
-      </Flex.Item>
-      <Icon icon="circle-outline" accessibilityLabel="unchecked circle" />
-    </Flex>
-    <Flex alignItems="center">
-      <Flex.Item flex="grow">
-        <Text>Saves</Text>
-      </Flex.Item>
-      <Icon icon="check-circle" color="default" accessibilityLabel="checked circle" />
-    </Flex>
-  </Flex>
-</Flex>
-`}
+            sandpackExample={
+              <SandpackExample
+                code={dontAlternative}
+                name="Don't - Alternative"
+                hideControls
+                hideEditor
+              />
+            }
           />
         </MainSection.Subsection>
       </MainSection>
+
       <AccessibilitySection
         name={generatedDocGen.displayName}
         description={`Switches should have [Labels](https://github.com/Label) that can be read by screen readers, and that can be clicked or tapped to make it easier for users to select and deselect. Make sure Label has an \`htmlFor\` prop that matches the \`id\` on the Switch. Test that the Switch and Label are properly connected by clicking or tapping on the label and confirming that it activates the Switch next to it.`}
@@ -198,121 +121,18 @@ function SwitchExample() {
         Whenever using Switch, always use a [Label](/web/label) with it to make your component accessible.`}
         >
           <MainSection.Card
-            defaultCode={`
-    function SwitchExample() {
-      const [switched1, setSwitched1] = React.useState(false);
-      const [switched2, setSwitched2] = React.useState(true);
-      const [switched3, setSwitched3] = React.useState(true);
-
-      return (
-        <Flex direction="column" gap={{ column: 2, row: 0 }} width="100%" maxWidth={300}>
-            <Heading size="300">Tune your home feed</Heading>
-            <Text>Turn off any boards that you don't want us to use for your home feed recommendations</Text>
-            <Flex direction="column" gap={{ column: 4, row: 0 }}>
-              <Flex alignItems="center">
-                <Flex.Item flex="grow">
-                  <Label htmlFor="boardfood">
-                    <Flex direction="column" gap={{ column: 1, row: 0 }}>
-                      <Text>Food</Text>
-                      <Text size="100">55 pins · 1 section</Text>
-                    </Flex>
-                  </Label>
-                </Flex.Item>
-                <Switch
-                  switched={switched1}
-                  onChange={() => setSwitched1(!switched1)}
-                  id="boardfood"
-                />
-              </Flex>
-              <Flex alignItems="center">
-                <Flex.Item flex="grow">
-                  <Label htmlFor="boardoutfits">
-                    <Flex direction="column" gap={{ column: 1, row: 0 }}>
-                      <Text>Outfits</Text>
-                      <Text size="100">138 pins · 5 sections</Text>
-                    </Flex>
-                  </Label>
-                </Flex.Item>
-                <Switch
-                  switched={switched2}
-                  onChange={() => setSwitched2(!switched2)}
-                  id="boardoutfits"
-                />
-              </Flex>
-              <Flex alignItems="center">
-                <Flex.Item flex="grow">
-                  <Label htmlFor="boardhomedecor">
-                    <Flex direction="column" gap={{ column: 1, row: 0 }}>
-                      <Text>Home Decor</Text>
-                      <Text size="100">33 pins · 2 sections</Text>
-                    </Flex>
-                  </Label>
-                </Flex.Item>
-                <Switch
-                  onChange={() => setSwitched3(!switched3)}
-                  id="boardhomedecor"
-                  switched={switched3}
-                />
-              </Flex>
-            </Flex>
-          </Flex>
-        );
-      }
-        `}
+            sandpackExample={<SandpackExample code={variantLabel} name="Variant - With a label" />}
           />
         </MainSection.Subsection>
         <MainSection.Subsection title="Disabled and switched combinations">
           <MainSection.Card
-            defaultCode={`
-function SwitchExample() {
-  return (
-    <Flex alignItems="center" gap={{ row: 8, column: 0 }}>
-      <Flex direction="column" gap={{ column: 2, row: 0 }}>
-        <Label htmlFor="base">
-          <Text>Base state</Text>
-        </Label>
-        <Switch
-          onChange={() => {}}
-          id="base"
-          switched={false}
-        />
-      </Flex>
-      <Flex direction="column" gap={{ column: 2, row: 0 }}>
-      <Label htmlFor="switched">
-        <Text>Switched</Text>
-      </Label>
-      <Switch
-        onChange={() => {}}
-        id="switched"
-        switched={true}
-      />
-    </Flex>
-    <Flex direction="column" gap={{ column: 2, row: 0 }}>
-        <Label htmlFor="disabled">
-          <Text>Disabled, not switched</Text>
-        </Label>
-        <Switch
-          onChange={() => {}}
-          id="disabled"
-          switched={false}
-          disabled={true}
-        />
-      </Flex>
-      <Flex direction="column" gap={{ column: 2, row: 0 }}>
-        <Label htmlFor="disabledAndSwitched">
-          <Text>Disabled and switched</Text>
-        </Label>
-        <Switch
-          onChange={() => {}}
-          id="disabledAndSwitched"
-          switched={true}
-          disabled={true}
-        />
-      </Flex>
-    </Flex>
-  );
-}
-    `}
+            sandpackExample={
+              <SandpackExample
+                code={variantCombinations}
+                name="Variant - Disabled and switched combinations"
+                layout="column"
+              />
+            }
           />
         </MainSection.Subsection>
       </MainSection>