+++ packages/gestalt/src/AvatarGroup.js
@@ -2,9 +2,9 @@
 import { forwardRef, type Node, type AbstractComponent, useCallback, useState } from 'react';
 import Box from './Box.js';
 import TapArea, { type OnTapType } from './TapArea.js';
-import AddCollaboratorsButton from './AvatarGroupAddCollaboratorsButton.js';
-import CollaboratorAvatar from './AvatarGroupCollaboratorAvatar.js';
-import CollaboratorsCount from './AvatarGroupCollaboratorsCount.js';
+import AddCollaboratorsButton from './AvatarGroup/AddCollaboratorsButton.js';
+import CollaboratorAvatar from './AvatarGroup/CollaboratorAvatar.js';
+import CollaboratorsCount from './AvatarGroup/CollaboratorsCount.js';
 import Flex from './Flex.js';
 
 const MAX_COLLABORATOR_AVATARS = 3;