+++ packages/gestalt/src/Controller.js
@@ -30,6 +30,7 @@ type OwnProps = {|
   rounding?: 2 | 4,
   shouldFocus?: boolean,
   size?: 'xs' | 'sm' | 'md' | 'lg' | 'xl' | number | null,
+  __dangerouslyIgnoreScrollBoundaryContainerSize?: boolean,
 |};
 
 type HookProps = {|
@@ -126,6 +127,7 @@ class Controller extends Component<Props, State> {
       rounding,
       shouldFocus,
       size,
+      __dangerouslyIgnoreScrollBoundaryContainerSize,
     } = this.props;
     const { relativeOffset, triggerBoundingRect } = this.state;
 
@@ -150,6 +152,9 @@ class Controller extends Component<Props, State> {
           shouldFocus={shouldFocus}
           triggerRect={triggerBoundingRect}
           width={width}
+          __dangerouslyIgnoreScrollBoundaryContainerSize={
+            __dangerouslyIgnoreScrollBoundaryContainerSize
+          }
         >
           {children}
         </Contents>